+++ src/OrchardCore/OrchardCore.Abstractions/Pooling/StringBuilderPool.cs
@@ -2,7 +2,7 @@ using System;
 using System.Diagnostics;
 using System.Text;
 
-namespace OrchardCore.Data.Pooling
+namespace OrchardCore.Abstractions.Pooling
 {
     /// <summary>
     /// The usage is:
