+++ src/OrchardCore/OrchardCore.Abstractions/Properties/AssemblyInfo.cs
@@ -1,0 +1,3 @@
+using System.Runtime.CompilerServices;
+
+[assembly: InternalsVisibleTo("OrchardCore.Data")]
