+++ src/OrchardCore/OrchardCore.Data/PoolingJsonContentSerializer.cs
@@ -1,0 +1,38 @@
+using System;
+using System.Buffers;
+using Newtonsoft.Json;
+using OrchardCore.Abstractions.Pooling;
+using YesSql;
+
+namespace OrchardCore.Data
+{
+    /// <summary>
+    /// Forwards to generic pooling JSON serializer with custom settings.
+    /// </summary>
+    internal sealed class PoolingJsonContentSerializer : IContentSerializer
+    {
+        private static readonly JsonSerializerSettings _jsonSettings = new JsonSerializerSettings
+        {
+            TypeNameHandling = TypeNameHandling.Auto,
+            DateTimeZoneHandling = DateTimeZoneHandling.Utc,
+            NullValueHandling = NullValueHandling.Ignore,
+            CheckAdditionalContent = false
+        };
+
+        private readonly PoolingJsonSerializer _inner;
+
+        public PoolingJsonContentSerializer(ArrayPool<char> arrayPool)
+        {
+            _inner = new PoolingJsonSerializer(arrayPool, _jsonSettings);
+        }
+
+        public object Deserialize(string content, Type type) => _inner.Deserialize(content, type);
+
+        public dynamic DeserializeDynamic(string content) => Deserialize<dynamic>(content);
+
+        // helps to resolve dynamic type
+        private object Deserialize<T>(string content) => _inner.Deserialize(content, typeof(T));
+
+        public string Serialize(object item) => _inner.Serialize(item);
+    }
+}
